split out signature.py
